Install RevPi-specific packages via apt-get
Up until now, RevPi-specific packages need to reside in debs-to-install/
and are installed with dpkg.  We're now able to download them with
apt-get instead, so leverage that.  The authenticity of the packages
is validated via the apt repo's signing key.  The functionality to
keep custom, image-specific packages in debs-to-install/ is retained.

Remove apache2 and libapache2-mod-php5 from the debs-to-download list
since they're pulled in by revpi-webstatus.

Signed-off-by: Lukas Wunner <lukas@wunner.de>