Prioritize wlan0 routes by default
If the RevPi Core has both a wired and a wireless connection and
receives a default route via DHCP on each, the wired default route
is currently preferred.  This is due to a policy hardcoded in dhcpcd,
as documented in the manpage:

    Metrics are used to prefer an interface over another one,
    lowest wins.  dhcpcd will supply a default metric of
    200 + if_nametoindex(3).
    An extra 100 will be added for wireless interfaces.

The rationale is probably that a wired connection is more reliable.
However on the RevPi Core, more often than not the wired connection is
a private LAN whereas the wireless connection is used to access the
Internet, so dhcpcd's policy isn't what the user wants.

Reduce the metric of wlan0 routes to avoid that users have to amend the
configuration for this relatively common use case.

Signed-off-by: Lukas Wunner <lukas@wunner.de>