Disable swap, gui & autologin
Disable swap by default to reduce eMMC write activity and increase free
disk space.  This reduces available memory by 100 MB, which however is
not expected to make a huge dent for the majority of users.

Disable GUI to improve boot time on the CM1 and conserve memory.  Most
RevPi Cores are likely to be installed in a cabinet and rarely accessed
via HDMI.

Disable autologin to prevent tampering with devices by anyone with
physical access.

Signed-off-by: Lukas Wunner <lukas@wunner.de>