Preserve disk id when creating shrunk image
Starting with Raspbian 2017-04-10, the Foundation no longer identifies
the boot and root partition in /etc/fstab as "/dev/mmcblk0", but by
the Raspbian image's disk id 0x402e4a57.  The rationale is that flashing
the image to a USB stick will now "just work":
https://www.raspberrypi.org/forums/viewtopic.php?f=66&t=180716

Unfortunately when we shrink the root partition using GNU parted, the
disk id is silently changed to a new value, resulting in boot failure
because systemd cannot remount the root partition specified in
/etc/fstab read/write.

Ensure the disk id is preserved when creating the shrunk image by
avoiding GNU parted altogether and relying only on sfdisk.

Signed-off-by: Lukas Wunner <lukas@wunner.de>