shrink_image: Cope with FAT partition of arbitrary size
Starting with Raspbian 2017-04-10, the Foundation has shrunk the FAT
partition from 63 MBytes to just 41 MBytes.  Because the size of the
ext4 partition is hardcoded, 22 MBytes are now unoccupied in our image.

Additionally, because the beginning of the ext4 partition is copied to
the shrunk image, newer versions of mkfs.ext4 issue a warning and ask
for confirmation.

Fix by retrieving the size of the FAT partition from the source image
and calculating the size of the ext4 partition to occupy the remainder
of the eMMC.

The geometry of the Foundation's partition table causes sfdisk to
complain, necessitating the "-f" force switch:

    sfdisk: end of partition 2 has impossible value for cylinders: 475 (should be in 0-474)

Reported-by: Riccardo Magliocchetti <riccardo@menodizero.it>
Signed-off-by: Lukas Wunner <lukas@wunner.de>