customize_image: Pivot to new PID namespace
customize_image.sh fails if apache is running on the host because the
apache postinst script run in the chroot incorrectly believes it was
unable to stop apache.

Fix by pivoting to a new pid namespace if newpid(1) is installed.
The pid is thus always 2 and cannot be used as a pseudo-random id for
temporary directories, so switch to mktemp instead.

Signed-off-by: Lukas Wunner <lukas@wunner.de>