customize_image: Peg CPU at 1200 MHz on CM3
Raspbian ships with a custom initscript /etc/init.d/raspi-config which
enables the "ondemand" cpu governor with a set of handcrafted parameters
unless the shift key is pressed on boot.

The ondemand governor downclocks the CPU to 600 MHz if CPU load doesn't
exceed 50%.  piControl usually stays below that level.  It has been
observed that UART communication and kthread wakeup latency becomes
jittery at 600 MHz.  It is unclear if the jitter is caused by the lower
CPU frequency or by occasional frequency changes.  Presumably when the
frequency is changed, the ARM core needs to be stopped briefly to change
the clock speed.

Moreover, spi0 throughput is degraded by 40% at 600 MHz because its
clock is derived from the CPU's clock. The Foundation is not inclined to
improve that:
https://github.com/raspberrypi/linux/issues/2094#issuecomment-312611868

Disable the raspi-config initscript by default, install cpufrequtils and
let it change the governor to "performance" on boot, thus pegging the
CPU at 1200 MHz (unless core temperature exceeds 80 °C).  Users can
always change the governor or disable /etc/init.d/cpufrequtils if this
default setting doesn't meet their needs.

Signed-off-by: Lukas Wunner <lukas@wunner.de>