customize_image: Disable procon-web-iot by default
The packages for logi-rts and other 3rd party software are currently
built with a "dh_systemd_enable --no-enable" override so that they're
not enabled by default on a freshly built image.

With procon-web-iot we've started to deviate from this rule so that on
a system where it was previously not installed, installing the package
manually results in the new service being enabled immediately.  This
however requires that the service is disabled when building the image.

Long term we'll move the other packages to the same scheme,
in conformance with common behavior on Debian.

Signed-off-by: Lukas Wunner <lukas@wunner.de>