treewide: Update for Raspbian 2018-11-13
Previously Raspbian was only available in two flavours, a lite image and
a full image.  The lite image is considerably smaller but lacks a
graphical desktop environment and many applications.

Starting with Raspbian 2018-11-13, a third flavour is available which
contains the desktop environment but not the extra applications.
Incidentally, the packages not included in this new flavour are largely
the same that we previously removed to fit Raspbian on a 3.9 GByte eMMC,
so we'll base our image on this third flavour going forward.

This Raspbian image is smaller than 3.9 GByte and needs to be grown
instead of shrunk.  Rework shrink_image.sh accordingly and rename to
resize_image.sh.  Drop packages which are gone in this image from
debs-to-remove.

Raspbian now ships with VLC pre-installed, which is nice but occupies
about 60 MBytes and doesn't seem particularly relevant for industrial
use cases, so remove it.

The package poppler-data (12 MBytes) can no longer be removed because
libgs9 depends on it.

The version of dhcpcd included with Raspbian now contains the fix
explained in commit bdfbddd88a30 ("customize_image: Bring up eth0 again
after USB bus reset (dhcpcd edition)") and commit 5ced866199b4
("customize_image: Update dhcpcd5 package"), so it no longer needs to be
upgraded explicitly upon image customization.

Signed-off-by: Lukas Wunner <lukas@wunner.de>