customize_image: Use chroot for installation of local packages
customize_image allows installation of local packages to the image.  The
user is expected to place the packages in imagebakery's debs-to-install
directory.  Because the packages need to be accessible upon installation,
they're installed with "dpkg --root" instead of "chroot dpkg", but that
fails if the packages have dependencies which are missing on the host
system.

Fix by chrooting properly and making the packages accessible in the
chroot environment using a bind mount.

Signed-off-by: Lukas Wunner <lukas@wunner.de>