resize_image: Drop in favor of resizing on first boot
The Raspbian image flavour we're now based on is smaller than 3.9 GByte,
hence fits on the eMMC of a CM1 and CM3 without the need for shrinking.

To avoid unused space on the eMMC, the image is currently grown to
3.9 GByte using resize_image.sh before customizing it.  However for
compatibility with the CM3+, the root filesystem will be grown anyway on
first boot using revpi-factory-reset, which renders resize_image.sh
obsolete.

Drop resize_image.sh.  A desirable side effect of not growing the image
up front is a speedup of about 1 minute when flashing the image.

Signed-off-by: Lukas Wunner <lukas@wunner.de>