customize_image: Shrink image after customization
Shrink the ext4 filesystem and its partition to the minimum size and
truncate the image file accordingly.  This speeds up flashing by more
than 2 minutes.  (Previously, flashing a full 3.9 GByte image took
roughly 10 minutes, now we're down to less than 8 minutes.)

The ext4 filesystem and partition will be grown to the maximum size on
first boot by revpi-factory-reset.  This takes just 1 or 2 seconds and
is necessary anyway to support all CM3+ variants with a single image.

Signed-off-by: Lukas Wunner <lukas@wunner.de>