customize_image: Work around ld.so warnings
The Foundation recently added the "raspi-copies-and-fills" package to
their apt repository.  It contains shared libraries with optimized
string functions for armv6 and armv7 CPUs.

The shared libraries are loaded by /etc/ld.so.preload via the $PLATFORM
token expansion (see ld.so(8)), which fails on qemu with a warning
message.  The warning itself is harmless, but it clutters up the screen
output when building an image and obscures real error messages.

At the behest of the Foundation's own Serge Schneider, future versions
of qemu will properly set the $PLATFORM value, thereby avoiding the
warning messages: https://bugs.launchpad.net/qemu/+bug/1813034

However it will take a while until an updated version of qemu is
released.  As a temporary quick fix, exclude this single package when
upgrading the distribution's packages.

Signed-off-by: Lukas Wunner <lukas@wunner.de>