templates/config.txt: Add [pi4] model filter to combine with [cm4s]
In the firmware for CM3, the [cm4s] is not recognized, which leads
the configurations after it will be applied. However, the [pi4] is
recognized and is seen as mismatch, thus the configurations after
that will not be applied as expected.

And in the firmware for CM4S, the [pi4] is also recognized and is
seen as match, thus the configurations after that will be applied
as expected.

By the way, as tested, the [cm4] is recognized on CM4S, but is
seen as mismatch, thus the configuration after that will not be
applied. This is not as expected, so it should not be used in this
case.

All above are gotten from testing with updated raspberrypi-bootloader
(1:1.20220308~buster-1)

Signed-off-by: Zhi Han <z.han@kunbus.com>