Add option minimize to build minimal debs to image
With the option -m, --minimize, the generated image will be a subset
of the image being built without it. In principle, the GUI based
packages are removed.

The packages in min-debs-to-download will primarily be installed, and
then if the -m or --minimize is not specified, the packages in
debs-to-download will also be installed subsequently.

Signed-off-by: Zhi Han <z.han@kunbus.com>