Remove NOP statements and fix apt.conf
Removes two uneccessary lines mangling with our sources.list.
The first removed line adds a raspbian mirror to our sources.list,
although raspbian is not used by the foundation since a while. In the
end it's not really relevant as the second line removes the previously
added repository. Furthermore the second line removes the first line of
/etc/apt/apt.conf which reverts the `APT::Install-Recommends` setting
from a few lines above.

Signed-off-by: Nicolai Buchwitz <n.buchwitz@kunbus.com>