install dt-blob.bin from revpi-dt-blob.dtbo
To support the HAT EEPROM, the VideoCore functionalities of ID_SDA
and ID_SCL need to set to pin 0 and pin 1 respectively.

For CM3 and CM4S, these are done in revpi-dt-blob.dtbo and by installing
it to dt-blob.bin the settings get applied during booting.

For the devices without HAT EEPROM, these settings should be overridden
by installing the device specific dt-blob in factory reset.

Signed-off-by: Zhi Han <z.han@kunbus.com>