Use fake procfs for NodeRed setup
NodeRed installer script relies on /proc/{mem,cpu}info in order to configure
the max memory footprint and other stuff. This will fail because there is no
procfs in the chroot environment and it does not make sense to mount the build
hosts procfs as their cpu architecture and amount of memory is highly likely to
be different.

See: https://github.com/node-red/linux-installers/blob/master/deb/update-nodejs-and-nodered#L612-L614

Fix this by creating a fake procfs, which contains only these two files, which
are copied from a RevPi Core S.

Signed-off-by: Nicolai Buchwitz <n.buchwitz@kunbus.com>