Claim extra space in the image for EOL tests
For the EOL test it is required to have some free space in the image in order
to copy files for provisioning. Previously, this was accomplished through the
use of magic (probably due to alignment or other implicit fs defaults).

This won't work with the latest RPi Buster image.

Fix this by explicitly claiming additional free space when shrinking the image.

Signed-off-by: Nicolai Buchwitz <n.buchwitz@kunbus.com>