Use blockdev for image size calculation
The current approach (parse output of parted) which we use to calculate the
size of our source image is likely to fail on systems with non-english locale
setups. This has been reported by Ramiro while working on the firstboot stuff.

Fix this by using blockdev (part of utils-linux) to get number of used blocks.

Reported-by: Ramiro Omar Gsponer <r.gsponer@kunbus.com>
Signed-off-by: Nicolai Buchwitz <n.buchwitz@kunbus.com>