Optional output image name
Add an optional parameter for the name of the output image. If this name is
given, the imagebakery will copy the source image to it and apply all
modifications. Otherwise, a warning message is displayed, which the user must
acknowledge before the program can continue. The modifications will be applied
in-place as before in the aforementioned case.

Signed-off-by: Nicolai Buchwitz <n.buchwitz@kunbus.com>