Node Red: Install RevPi nodes from git release
The RevPi Nodes only work with a specific version of the backend server.
This can lead to incompatible setups, when a new version is published to
npm and we do not update our backend server too.

Fix this, until there is a better solution, by installing the RevPi Nodes
from our git repository and use a specific release tag.

Signed-off-by: Nicolai Buchwitz <n.buchwitz@kunbus.com>