Update exclude list for Bullseye images
revpi-tools is now available as a src package, so we don't need to exclude it
from the source collection. The same applies to revpimodio2 which is available
as source package too.

New to the exclude list is the package 'widevine' which was introduced in the
Raspberry Pi OS as a dependency for the chromium package.
See pi-gen commit 69cd27944c9d81b8170e8ade5ccbb23895599ed1

Signed-off-by: Nicolai Buchwitz <n.buchwitz@kunbus.com>