Revert "Enable kms-v3d as default"
This reverts commit e9d4274039df69a66b95c9bdc54e5e78e1c5db68.

Loading an overlay in the config.txt is a bad idea. We can not package
the confitg.txt. This will lead to configuration fragmentation. We
should handle these changes in our dt overlays.

As we don't have any device with more then one HDMI, there is no need to
set max_framebuffers=2.

https://www.raspberrypi.com/documentation/computers/config_txt.html#max_framebuffers