fix: Ensure that pileft and piright are brought up
The ethernet interfaces pileft and piright, which are used for gateway
communication need to be up, so piControl can initialize the communication.

Without that fix the interfaces stay down and gateways cannot be used. On older
devices without attached irq pins the pi(left|right)-quirks.service ensures
that the interfaces will be disabled if not gateway is attached (part of
revpi-tools).

Signed-off-by: Nicolai Buchwitz <n.buchwitz@kunbus.com>