fix: Avoid writing logrotate configuration
The configuration for logrotate should be done in separate configuration
files in `/etc/logrotate.d/` and should ideally live in a separate
package so that the configuration can be updated for existing systems if
necessary or removed by the user if it is not needed.

Instead, the configuration now lives in the `revpi-base-files` package.

Signed-off-by: Thomas Böhler <t.boehler@kunbus.com>