feat: Mount host filesystem in interactive mode
For the interactive mode of install_debs_into_image, a chroot is
performed into the image. If required files are to be transferred from
the host system to the image or scripts are to be executed, you need
access to the file system from the host. The current working directory
is now mounted in the image in interactive mode (-i) in /mnt.

Signed-off-by: Sven Sager <s.sager@kunbus.com>