feat: Add script so shrink custom images
To increase the performance when flashing, an image should be kept as
small as possible. If the image of a Revolution Pi is created, it can
cover the full size of the eMMC memory. This script reduces the size of
the partition and ensures that the "firstboot.service" is executed at
the first start, which expands the image back to the full size of the
eMMC.

Signed-off-by: Sven Sager <s.sager@kunbus.com>